// postgre-server/prisma/generated/zod/inputTypeSchemas/TagsDocumentWhereInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { StringFilterSchema } from './StringFilterSchema';
import { DateTimeFilterSchema } from './DateTimeFilterSchema';
import { TagRelationFilterSchema } from './TagRelationFilterSchema';
import { TagWhereInputSchema } from './TagWhereInputSchema';
import { DocumentRelationFilterSchema } from './DocumentRelationFilterSchema';
import { DocumentWhereInputSchema } from './DocumentWhereInputSchema';

export const TagsDocumentWhereInputSchema: z.ZodType<Prisma.TagsDocumentWhereInput> = z.object({
  AND: z.union([ z.lazy(() => TagsDocumentWhereInputSchema),z.lazy(() => TagsDocumentWhereInputSchema).array() ]).optional(),
  OR: z.lazy(() => TagsDocumentWhereInputSchema).array().optional(),
  NOT: z.union([ z.lazy(() => TagsDocumentWhereInputSchema),z.lazy(() => TagsDocumentWhereInputSchema).array() ]).optional(),
  tagId: z.union([ z.lazy(() => StringFilterSchema),z.string() ]).optional(),
  documentId: z.union([ z.lazy(() => StringFilterSchema),z.string() ]).optional(),
  createdAt: z.union([ z.lazy(() => DateTimeFilterSchema),z.coerce.date() ]).optional(),
  updatedAt: z.union([ z.lazy(() => DateTimeFilterSchema),z.coerce.date() ]).optional(),
  Tag: z.union([ z.lazy(() => TagRelationFilterSchema),z.lazy(() => TagWhereInputSchema) ]).optional(),
  Document: z.union([ z.lazy(() => DocumentRelationFilterSchema),z.lazy(() => DocumentWhereInputSchema) ]).optional(),
}).strict();

export default TagsDocumentWhereInputSchema;
